refactor: move extract_client_info to http.utils and update imports
