+++ src/audit/service.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from fastapi import Depends, Request
+from fastapi import Depends
 from loguru import logger
 from sqlalchemy.ext.asyncio import AsyncSession
 
@@ -77,13 +77,3 @@ async def get_audit_service(
 ) -> AuditService:
     repository = AuditRepository(session=session)
     return AuditService(repository=repository)
-
-
-def extract_client_info(request: Request) -> tuple[str | None, str | None]:
-    user_agent = request.headers.get("user-agent")
-    forwarded_for = request.headers.get("x-forwarded-for")
-    if forwarded_for:
-        ip = forwarded_for.split(",")[0].strip()
-    else:
-        ip = request.client.host if request.client else None
-    return user_agent, ip
