+++ src/http/utils.py
@@ -1,0 +1,11 @@
+from fastapi import Request
+
+
+def extract_client_info(request: Request) -> tuple[str | None, str | None]:
+    user_agent = request.headers.get("user-agent")
+    forwarded_for = request.headers.get("x-forwarded-for")
+    if forwarded_for:
+        ip = forwarded_for.split(",")[0].strip()
+    else:
+        ip = request.client.host if request.client else None
+    return user_agent, ip
